🚚 ➕ ♻️ Move mongo code into separate private lib
