+++ libs/mongo/.babelrc.json
@@ -1,0 +1,17 @@
+{  
+  "presets": [
+    "@babel/preset-env",
+    "@babel/preset-typescript"
+  ],
+  
+  "plugins": [
+    [
+      "module-resolver",
+      {
+        "root": ["."],
+        "alias": {}
+      }
+    ],
+    "@babel/plugin-transform-runtime"
+  ]
+}
