+++ libs/mongo/README.md
@@ -1,0 +1,11 @@
+# `mongo`
+
+> TODO: description
+
+## Usage
+
+```typescript
+import lib from '@discord-bot/mongo'
+
+// TODO: DEMONSTRATE API
+```
