+++ libs/mongo/jest.config.js
@@ -1,0 +1,15 @@
+const path = require("path")
+
+module.exports = {
+  name: 'lib-test',
+  displayName: {
+    name: 'LIB: LIB-TEST',
+    color: 'cyan',
+  },
+
+  moduleDirectories: ["node_modules", path.join(__dirname, "src")],
+
+  testPathIgnorePatterns: [
+    "dist/",
+  ],
+}
