+++ libs/mongo/src/Utils/dropAllCollections.ts
@@ -1,16 +1,5 @@
+import logger from "./logger"
 import { Mongoose } from "mongoose"
-import logger from "../logger"
-
-// utility function to remove all documents from all collections
-export const deleteAllDocuments = (conn: Mongoose) => {
-  // get all collections
-  const collections = Object.values(conn?.connection.collections ?? {})
-  logger.debug("Utils/Mongo: Deleting all documents from all collections")
-  // for all collections, remove all documents
-  return Promise.all(collections
-    .map(collection => collection.deleteMany({}))
-  )
-}
 
 // utility function to drop all collections
 export const dropAllCollections = (conn: Mongoose) => {
