+++ libs/mongo/src/Utils/logger.ts
@@ -1,0 +1,5 @@
+import { createLogger } from "@discord-bot/logging"
+
+export default createLogger({
+  logLevel: "none",
+})
