+++ libs/mongo/src/index.ts
@@ -1,0 +1,7 @@
+export * from "./connect"
+
+// export mongoose and mongodb
+export * as mongoose from "mongoose"
+export * as mongodb from "mongodb"
+
+export * from "./Utils"
