+++ libs/mongo/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.json",
+  
+  "compilerOptions": {
+    "baseUrl": "src",
+    "rootDir": "./src",
+    "outDir": "./dist",
+  },
+}
