+++ modules/fantasy-battle/src/Commands/bio.test.ts
@@ -2,7 +2,7 @@ import { mockMessage } from "@discord-bot/discord-mock"
 import PcModel from "Models/PlayerCharacter"
 import PlayerUserModel from "Models/PlayerUser"
 import { mockDmMessage, mockPlayerMessage } from "Utils/Mock/mockMessage"
-import { useDbConnection } from "Utils/Mongo/mongoTest"
+import { useDbConnection } from "@discord-bot/mongo"
 
 import { test, execute } from "./bio"
