+++ modules/fantasy-battle/src/Commands/changeResource.test.ts
@@ -3,7 +3,7 @@ import PcModel from "Models/PlayerCharacter"
 
 import PlayerUserModel from "Models/PlayerUser"
 import { mockDmMessage, mockPlayerMessage } from "Utils/Mock/mockMessage"
-import { useDbConnection } from "Utils/Mongo/mongoTest"
+import { useDbConnection } from "@discord-bot/mongo"
 
 import { test, execute } from "./changeResource"
