+++ modules/fantasy-battle/src/Commands/rollDmg.test.ts
@@ -2,7 +2,7 @@ import { mockMessage } from "@discord-bot/discord-mock"
 import PcModel from "Models/PlayerCharacter"
 import PlayerUserModel from "Models/PlayerUser"
 import { mockDmMessage, mockPlayerMessage } from "Utils/Mock/mockMessage"
-import { useDbConnection } from "Utils/Mongo/mongoTest"
+import { useDbConnection } from "@discord-bot/mongo"
 
 import { test, execute } from "./rollDmg"
 
@@ -75,7 +75,7 @@ describe("Command: rollDmg", () => {
 
         const playerUser = PlayerUserModel.createUser({ userId: message.author.id, username: message.author.username })
         playerUser.addCharacter(PcModel.createCharacter({ name: "Horu" }))
-        playerUser.characters[0].attributes.Might.value = 4
+        playerUser.activeChar.attributes.Might.value = 4
         await playerUser.save()
         await execute(message, test.exec(message.content)!)
