+++ modules/fantasy-battle/src/Models/PlayerCharacter/virtuals/highestPhysical.test.ts
@@ -1,4 +1,4 @@
-import { useDbConnection } from "Utils/Mongo/mongoTest"
+import { useDbConnection } from "@discord-bot/mongo"
 import PcModel from ".."
 
 import highestPhysical from "./highestPhysical"
