+++ modules/fantasy-battle/src/Models/PlayerUser/statics/create.test.ts
@@ -1,4 +1,4 @@
-import { useDbConnection } from "Utils/Mongo/mongoTest"
+import { useDbConnection } from "@discord-bot/mongo"
 
 import create, { createPlayerUserProps } from "./create"
 import PlayerUserModel from ".."
